fix: Enable selection of currently reading
In the search page, when selecting currently reading it was not
triggering the handle selection event because the listener was for
onChange, and the default value is currently reading.  So add another
event for onClick.

Now need to update state of book in search page to match that of book
shelves.